backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
firebase-admin==6.2.0
Pillow==10.0.0
reportlab==4.0.4
python-dotenv==1.0.0
pandas==2.0.3
APScheduler==3.10.1
PyJWT==2.8.0
matplotlib==3.7.2
